REFACTOR: change document to model module in presentation layer
