+++ user/presentation/build.gradle.kts
@@ -16,6 +16,7 @@ dependencies {
     implementation(project(":user:core"))
     implementation(project(":user:persistence"))
     implementation(project(":kafka"))
+    implementation(project(":model"))
 
 
     //WEBFLUX
